Validate categoryTitle

+++ src/main/java/ntnu/idatt1002/controllers/EditCategoryController.java
@@ -2,6 +2,7 @@ package ntnu.idatt1002.controllers;
 
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
+import javafx.scene.control.Label;
 import javafx.scene.control.TextField;
 import ntnu.idatt1002.service.CategoryService;
 import ntnu.idatt1002.service.TaskService;
@@ -12,6 +13,7 @@ import java.io.IOException;
 public class EditCategoryController {
 
     @FXML private TextField titleTextField;
+    @FXML private Label errorMessage;
 
     /**
      * Cancel button loads the tasks page back into center-content of dashboard
@@ -23,19 +25,24 @@ public class EditCategoryController {
     }
 
     public void buttonEditCategory(ActionEvent event) throws IOException {
-        // Make new category
-        CategoryService.addCategoryToCurrentUser(titleTextField.getText());
+        if(CategoryService.validateCategoryTitleSyntax(titleTextField.getText())){
+            // Make new category
+            CategoryService.addCategoryToCurrentUser(titleTextField.getText());
 
-        // Move tasks in old category to new category
-        TaskService.editCategoryOfTasks(TaskService.getTasksByCategory(UserStateService.getCurrentUserCategory()), titleTextField.getText());
+            // Move tasks in old category to new category
+            TaskService.editCategoryOfTasks(TaskService.getTasksByCategory(UserStateService.getCurrentUserCategory()), titleTextField.getText());
 
-        // Delete old category
-        CategoryService.deleteCategoryCurrentUser(UserStateService.getCurrentUserCategory());
+            // Delete old category
+            CategoryService.deleteCategoryCurrentUser(UserStateService.getCurrentUserCategory());
 
-        // Set current category to new one
-        UserStateService.setCurrentUserCategory(titleTextField.getText());
+            // Set current category to new one
+            UserStateService.setCurrentUserCategory(titleTextField.getText());
 
-        // Load dashboard into maincontent
-        DashboardController.getInstance().initialize();
+            // Load dashboard into maincontent
+            DashboardController.getInstance().initialize();
+        } else {
+            errorMessage.setText("Title need to be between 0 and 30 characters");
+        }
+
     }
 }
